requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langchain-huggingface>=0.1.0
langchain-qdrant>=0.1.0
qdrant-client>=1.6.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
torch>=2.0.0
pydantic>=2.5.0 
langchain-postgres